This is the new Pull request to merge the pypi-install work

+++ conda/lock.py
@@ -13,10 +13,12 @@ the race condition-proof os.makedirs.
 For now, there is one global lock for all of conda, because some things happen
 globally (such as downloading packages).
 
+We don't raise an error if the lock is named with the current PID
 """
 
 import os
-from os.path import exists, join
+from os.path import join
+import glob
 
 
 LOCKFN = '.conda_lock'
@@ -28,10 +30,14 @@ class Locked(object):
     """
     def __init__(self, path):
         self.path = path
-        self.lock_path = join(self.path, LOCKFN)
+        self.end = "-" + str(os.getpid())
+        self.lock_path = join(self.path, LOCKFN + self.end)
+        self.pattern = join(self.path, LOCKFN + '-*')
+        self.remove = True
 
     def __enter__(self):
-        if exists(self.lock_path):
+        files = glob.glob(self.pattern)
+        if files and not files[0].endswith(self.end):
             # Keep the string "LOCKERROR" in this string so that external
             # programs can look for it.
             raise RuntimeError("""\
@@ -40,14 +46,18 @@ The lock %s was found. Wait for it to finish before continuing.
 If you are sure that conda is not running, remove it and try again.
 You can also use: $ conda clean --lock""" % self.lock_path)
 
-        try:
-            os.makedirs(self.lock_path)
-        except OSError:
-            pass
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        for path in self.lock_path, self.path:
+        if not files:
             try:
-                os.rmdir(path)
+                os.makedirs(self.lock_path)
             except OSError:
                 pass
+        else: # PID lock already here --- someone else will remove it.
+            self.remove = False
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        if self.remove:
+            for path in self.lock_path, self.path:
+                try:
+                    os.rmdir(path)
+                except OSError:
+                    pass
